UBUNTU: [config] linux-raspi2: Fix config-check errors
BugLink: https://bugs.launchpad.net/bugs/1861724

We started to enforce some directives from the annotations file that
were wrongly being ignored before, this behavior was introduced by the
commit 1e85cdaf32d7 ("UBUNTU: [Packaging]: config-check: some
enforcement is ignored").

For linux-raspi2, ignore the VFIO changes since they are not supported
for raspi and update the remaining ones to match linux-generic.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>